fix inverse trig functions; associate radix constants
fix copy-paste typo on inverse trig functions where approximation would silently fail.

implement `unreachable!()` check in `expr::cast` which will panic in the event of similar errors in the future (that is, when the given floating-point fuction returns a non-finite result).

change namespace of radix constants from the `radix` module to the `Radix` struct.